refactors user integration tests into a user_flows test. Adds flows integration tests for runcats, speedruns, and games.
