+++ test/integration/game_flows_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class GameFlowsTest < ActionDispatch::IntegrationTest
+  # test "the truth" do
+  #   assert true
+  # end
+end
