+++ test/integration/runcat_flows_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class RuncatFlowsTest < ActionDispatch::IntegrationTest
+  # test "the truth" do
+  #   assert true
+  # end
+end
