+++ test/integration/speedrun_flows_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class SpeedrunFlowsTest < ActionDispatch::IntegrationTest
+  # test "the truth" do
+  #   assert true
+  # end
+end
