--- test/integration/users_profile_test.rb
@@ -1,17 +1,0 @@
-require 'test_helper'
-
-class UsersProfileTest < ActionDispatch::IntegrationTest
-    include ApplicationHelper
-
-    def setup
-        @user = users(:michael)
-    end
-
-    test "profile display" do
-        get user_path(@user)
-        assert_template 'users/show'
-        assert_select 'title', full_title(@user.name)
-        assert_select 'h1', text: @user.name
-        assert_select 'h1>img.gravatar'
-    end
-end
