--- test/integration/users_signup_test.rb
@@ -1,22 +1,0 @@
-require 'test_helper'
-
-class UsersSignupTest < ActionDispatch::IntegrationTest
-
-    test "invalid signup information" do
-        get signup_path
-        assert_no_difference 'User.count' do
-            post users_path, params: {user: {name: "", email: "user@invalid", password: "foo", password_confirmation: "bar"}}
-        end
-        assert_template 'users/new'
-    end
-
-    test "valid signup information" do
-        get signup_path
-        assert_difference 'User.count', 1 do
-            post users_path, params: {user: {name: "Example User", email: "user@example.com", password: "password", password_confirmation: "password"}}
-        end
-        user = assigns(:user)
-        log_in_as(user)
-        assert is_logged_in?
-    end
-end
